connect to app

+++ SafePoint.Data/Entities/Shelter.cs
@@ -10,8 +10,8 @@ namespace SafePoint.Data.Entities
 
         public string Name { get; set; }
 
-        public float Lat { get; set; }
+        public decimal Lat { get; set; }
 
-        public float Lon { get; set; }
+        public decimal Lon { get; set; }
     }
 }
